Make more configurable.
In joymapper, let the user specify a different controller device than
/dev/input/js0.

In midimapper, let the user specify the name of the MIDI controller
they want to communicate with.  Clarify the port-wrangling code.  As
before, we still default to the first device, which is what ALSA (at
least, the alsa_midi library) considers the most usable.

I don't happen to have the controller that Dirk uses, so I haven't
tested all of that code yet.